Lint cleanup, begin testing in different files.
